+++ frontend/src/components/TableComplete/index.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from "react";
+import React, { useEffect, useState } from "react";
 import { Box, Button, Typography } from "@mui/material";
 import { Renderer } from "components/Renderer";
 import { Table } from "components/Table";
@@ -60,31 +60,21 @@ export const TableComplete = ({
     currentRowData,
   } = useFormControls();
 
-  const { handlePost, handleUpdate, Notification } = useFormSubmit();
+  const { handlePost, handleUpdate, handleDelete, Notification } = useFormSubmit();
 
   const [userCapabilities, setUserCapabilities] = useState<string[]>([]);
 
   const { axiosAll } = useAxios();
 
-  /**
-   * returns destructured props from the useFormatTableData hook.
-   *
-   * @param   {string}   tableName   - The name of the table that you are wanting data from.
-   * @param   {string}   apiEndPoint - The enpoint as which the API query will use for it's call.
-   * @param   {Function} handleClick - Function passed to the "view" button of the Table component.
-   * @returns {object}               {data, isLoading}  - "data" contains the columns and rows of data for your table.  isLoading is a boolean prop that changes to true if quering data and false if it has received the data.
-   */
-
-  const { data, isLoading } = useFormatTableData({
-    tableName: tableName,
-    apiEndPoint: url.getAll,
-    columnWidths: columnWidths,
-    handleClick: handleOpen,
-  });
-
-  useEffect(() => {
-    setUserCapabilities(data?.user?.capabilities);
-  }, [data]);
+  const hasRole = (requiredRole: string) => {
+    let allowed = false;
+    if (Array.isArray(userCapabilities) && userCapabilities.length > 0) {
+      /* eslint "no-warning-comments": [1, { "terms": ["todo", "fixme"] }] */
+      // todo: This or needs to check for if user has pmo-sys-admin as well, but for testing this should be fine.
+      allowed = userCapabilities.includes(requiredRole) || "users_update_one" === requiredRole;
+    }
+    return allowed;
+  };
 
   /**
    * getApiData is the fetch function for react query to leverage.
@@ -113,15 +103,36 @@ export const TableComplete = ({
     return "";
   };
 
-  const hasRole = (requiredRole: string) => {
-    let allowed = false;
-    if (Array.isArray(userCapabilities) && userCapabilities.length > 0) {
-      /* eslint "no-warning-comments": [1, { "terms": ["todo", "fixme"] }] */
-      // todo: This or needs to check for if user has pmo-sys-admin as well, but for testing this should be fine.
-      allowed = userCapabilities.includes(requiredRole) || "users_update_one" === requiredRole;
+  const handleDeleteClick = () => {
+    if (!url.deleteOne) {
+      return false;
     }
-    return allowed;
+    handleDelete({
+      apiUrl: getApiUrl(url.deleteOne as string, currentRowData.id),
+      queryKeys: [url.getAll],
+      successMessage: "Deleted successfully.",
+      errorMessage: "There was an issue deleting the item.",
+    });
   };
+
+  /**
+   * returns destructured props from the useFormatTableData hook.
+   *
+   * @param   {string}   tableName   - The name of the table that you are wanting data from.
+   * @param   {string}   apiEndPoint - The enpoint as which the API query will use for it's call.
+   * @param   {Function} handleClick - Function passed to the "view" button of the Table component.
+   * @returns {object}               {data, isLoading}  - "data" contains the columns and rows of data for your table.  isLoading is a boolean prop that changes to true if quering data and false if it has received the data.
+   */
+  const { data, isLoading } = useFormatTableData({
+    tableName: tableName,
+    apiEndPoint: url.getAll,
+    columnWidths: columnWidths,
+    handleClick: handleOpen,
+  });
+
+  useEffect(() => {
+    setUserCapabilities(data?.user?.capabilities);
+  }, [data]);
 
   /**
    * used for the react query.
@@ -142,6 +153,7 @@ export const TableComplete = ({
       staleTime: Infinity,
     }
   );
+
   return (
     <>
       {hasRole(roles.get) && (
@@ -166,18 +178,17 @@ export const TableComplete = ({
                   }
                 }}
               />
-              <Box
-                m={1}
-                display="flex"
-                justifyContent="flex-end"
-                alignItems="flex-end"
-                onClick={() => {
-                  handleOpen();
-                  handleEditMode(true);
-                  handleFormType("new");
-                }}
-              >
-                {hasRole(roles.add) && <Button variant="contained">{`Add New ${itemName}`}</Button>}
+              <Box m={1} display="flex" justifyContent="flex-end" alignItems="flex-end">
+                {hasRole(roles.add) && (
+                  <Button
+                    onClick={() => {
+                      handleOpen();
+                      handleEditMode(true);
+                      handleFormType("new");
+                    }}
+                    variant="contained"
+                  >{`Add New ${itemName}`}</Button>
+                )}
               </Box>
             </>
           }
@@ -190,6 +201,8 @@ export const TableComplete = ({
         handleEditMode={handleEditMode}
         editMode={editMode}
         allowEdit={hasRole(roles.update)}
+        allowDelete={hasRole(roles.delete) && undefined !== url.deleteOne && "new" !== formType}
+        handleDelete={handleDeleteClick}
         handleFormType={handleFormType}
       >
         <>